add props test
